tests/wpt/mozilla/tests/input-events/dblclick-events-fired-every-two-clicks.html


<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>dblclick event is fired for every two mouse click events</title>
    <link rel="author" title="Martin Robinson" href="mailto:mrobinson@igalia.com" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/resources/testdriver.js"></script>
    <script src="/resources/testdriver-actions.js"></script>
    <script src="/resources/testdriver-vendor.js"></script>
    <style>
    #target
    {
        background-color: green;
        width: 200px;
        height: 200px;
    }
    </style>
  </head>
  <body>
    <div id="target"></div>
    <script>
        /// This tests that "dblclick" is fired for every two mouse events and that
        /// the detail is always 2, no matter how many clicks are fired.
        promise_test(async (t) => {
            const target = document.getElementById("target");
            const event_watcher = new EventWatcher(t, target, ["click", "dblclick"]);

            const actions_promise = new test_driver.Actions()
              .pointerMove(0, 0, {origin: target})
              .pointerDown().pointerUp()
              .pointerDown().pointerUp()
              .pointerDown().pointerUp()
              .pointerDown().pointerUp()
              .pointerDown().pointerUp()
              .pointerDown().pointerUp()
              .send();

            // Make sure the test finishes after all the input actions are completed.
            t.add_cleanup(() => actions_promise);

            const event = await event_watcher.wait_for([
                // First click.
                "click",
                // Second click.
                "click", "dblclick",
                // Third click.
                "click",
                // Fourth click.
                "click", "dblclick",
                // Fifth click.
                "click",
                // Sixth click.
                "click", "dblclick"
            ]);
            assert_equals(event.type, "dblclick");
            assert_equals(event.target, target);
            assert_equals(event.detail, 2);
        });
    </script>
  </body>
</html>
